Move inner data to lib.rs
