+++ Cargo.toml
@@ -18,9 +18,6 @@ exclude = ["/.*"]
 default = ["std"]
 std = ["parking"]
 
-# Unstable, test only feature. Do not enable this.
-__test = []
-
 [dependencies]
 crossbeam-utils = { version = "0.8.12", default-features = false }
 parking = { version = "2.0.0", optional = true }
